Why does it hurt when I pee?

+++ sdev/include/shaiya/include/RevengeMark.h
@@ -2,16 +2,14 @@
 #include <map>
 #include <vector>
 
-#include <shaiya/include/common.h>
-
 namespace shaiya
 {
     using CharId = unsigned long;
 
     struct RevengeMark
     {
-        ULONG killerId;
-        UINT32 killCount;
+        uint32_t killerId;
+        uint32_t killCount;
     };
 
     inline std::map<CharId, std::vector<RevengeMark>> g_revengeMark;
